Implemented support for newline characters as delimiters.

+++ StringCalculator.java
@@ -10,11 +10,11 @@ public class StringCalculator {
         if (numbers == null || numbers.isEmpty()) {
             return 0;
         }
-
+    
         String[] parts = numbers.split(",");
         int sum = 0;
         for (String part : parts) {
-            sum += Double.parseDouble(part.trim());
+            sum += Integer.parseInt(part.trim());
         }
         return sum;
     }
